Allow "partial" matches to the exclusion list
This fixes a regression introduced in
aa145a95ffa45c4bd026adf712d1ba1fdea32733:
If the clone ID that we look up in the exclusion list contains "-" or "0"
characters, we need to check only the other characters. This is now handled
by introducing a SimilaritySet class that uses exact lookups as long as
there are no '-' or '0', but falls back to the old algorithm otherwise.