Filter out some additional cloneIDs
One of the tasks of the compute_cells function is to remove partial cloneIDs
that are too short, but the function assumes that the partial cloneIDs
either start or end with a stretch of "-" characters.

So a cloneID such as "ATGT------------------------AC" would actually be kept
although it definitely does not contain enough nucleotides.