Filter out cloneIDs with deletions
We’ve always attempted to somehow use cloneIDs with deletions (deleted bases
are encoded with a '0' character), but I’m now convinced this introduces
more problems than it solves. In particular, where exactly the deletion is
placed is in principle unknown because the reference is unknown (just a
stretch of "N" bases). That is, cloneID 00ACGT could just as well be
AC00GT or ACGT00. Which of those versions is found depends on the aligner
and can be considered random.

There is one example in our test data, where we have a cell with cloneIDs
00TGTCAATCGTTCGGTTGAGCAAGATCTA and
TGTCAATCGTTCGGTTGAGCAAGATCTTAG. The string comparisons we do for
clustering and cloneID correction don’t see that one is a shifted version of
the other, so they consider them to be two different strings. These should
instead have been morgede into a single molecule.