Rename Clone to CellSet
The class is used to represent multiple cells, which is kind of what a clone
is, but we only use it to compress cells with identical sets of cloneIDs to
a single node. We don’t actually return Clone instances when computing
clones (these are instead represented as a list of cells). So to avoid
confusion, rename this to CellSet.